+++ bpf/insn/samples/trace_printk.c
@@ -1,6 +1,10 @@
+/**
+ * int bpf_trace_printk(const char *fmt, int fmt_size, ...);
+ */
 #include "bpf_insn_samples.h"
 
 static struct bpf_insn insns[] = {
+	/* R10: frame pointer */
 	BPF_MOV64_IMM(BPF_REG_1, 0x21),        /* '!' */
 	BPF_STX_MEM(BPF_H, BPF_REG_10, BPF_REG_1, -4),
 	BPF_MOV64_IMM(BPF_REG_1, 0x646c726f),   /* 'orld' */
@@ -13,7 +17,7 @@ static struct bpf_insn insns[] = {
 	BPF_STX_MEM(BPF_B, BPF_REG_10, BPF_REG_1, -2),
 	BPF_MOV64_REG(BPF_REG_1, BPF_REG_10),
 	BPF_ALU64_IMM(BPF_ADD, BPF_REG_1, -16),
-	BPF_MOV64_IMM(BPF_REG_2, 15),
+	BPF_MOV64_IMM(BPF_REG_2, 0xf), /* fmt_size=15 */
 	BPF_RAW_INSN(BPF_JMP|BPF_CALL, 0, 0, 0, BPF_FUNC_trace_printk),
 	BPF_MOV64_IMM(BPF_REG_0, 0),
 	BPF_EXIT_INSN(),