+++ bpf/kfunc/samples/bpf_trace_printk.bpf.c
@@ -1,0 +1,13 @@
+/**
+ * int bpf_trace_printk(const char *fmt, int fmt_size, ...);
+ *
+ * linux v4.0-rc5-419-g9c959c863f82
+ * commit 9c959c863f82 ("tracing: Allow BPF programs to call bpf_trace_printk()")
+ */
+#include "vmlinux.h"
+
+void test(void)
+{
+	char s1[] = "Hello, World!";
+	bpf_trace_printk(s1, 15);
+}