+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to
 - profile.d: add `make_tl.sh` to alias system make command;
 - template: include pure header for main.mk, etc.;
 #### Added
-- bpf/kfunc: add `bpf_strncmp()` test;
+- bpf/kfunc: add `bpf_strncmp()`/`bpf_trace_printk()` test;
 - syscall: add empty fadvise(2) to test;
 - fs: add f2fs, lustre;
 - bpf/kfunc: add samples;